CIV-10391 changes
