+++ src/main/java/uk/gov/hmcts/reform/civil/callback/CaseEvent.java
@@ -222,7 +222,6 @@ public enum CaseEvent {
     GENERATE_CLAIM_FORM(CAMUNDA),
     GENERATE_CLAIM_FORM_SPEC(CAMUNDA),
 
-    SET_LIP_RESPONDENT_RESPONSE_DEADLINE(CAMUNDA),
     PROCESS_FULL_DEFENCE(CAMUNDA),
     PROCESS_FULL_DEFENCE_SPEC(CAMUNDA),
     PROCESS_PART_ADMISSION_DEFENCE_SPEC(CAMUNDA),
