+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/CreateClaimSpecAfterPaymentCallbackHandler.java
@@ -2,7 +2,6 @@ package uk.gov.hmcts.reform.civil.handler.callback.user;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
@@ -12,9 +11,12 @@ import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
 import uk.gov.hmcts.reform.civil.model.BusinessProcess;
 import uk.gov.hmcts.reform.civil.model.CaseData;
+import uk.gov.hmcts.reform.civil.service.DeadlinesCalculator;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 import static java.util.Collections.singletonList;
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
@@ -22,7 +24,6 @@ import static uk.gov.hmcts.reform.civil.callback.CallbackType.SUBMITTED;
 import static uk.gov.hmcts.reform.civil.callback.CallbackVersion.V_1;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.CREATE_CLAIM_SPEC_AFTER_PAYMENT;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class CreateClaimSpecAfterPaymentCallbackHandler extends CallbackHandler {
@@ -30,6 +31,7 @@ public class CreateClaimSpecAfterPaymentCallbackHandler extends CallbackHandler 
     // This should be created by the handler/service that receives payment info via our endpoint.
     private static final List<CaseEvent> EVENTS = singletonList(CREATE_CLAIM_SPEC_AFTER_PAYMENT);
     private final ObjectMapper objectMapper;
+    private final DeadlinesCalculator deadlinesCalculator;
 
     @Override
     protected Map<String, Callback> callbacks() {
@@ -45,14 +47,16 @@ public class CreateClaimSpecAfterPaymentCallbackHandler extends CallbackHandler 
     }
 
     private CallbackResponse changeStateToCaseIssued(CallbackParams callbackParams) {
-        Long caseId = callbackParams.getCaseData().getCcdCaseReference();
         CaseData caseData = callbackParams.getCaseData();
         CaseData.CaseDataBuilder dataBuilder = caseData.toBuilder();
         dataBuilder.businessProcess(BusinessProcess.ready(CREATE_CLAIM_SPEC_AFTER_PAYMENT));
-
+        if (Objects.isNull(caseData.getRespondent1ResponseDeadline())) {
+            dataBuilder.respondent1ResponseDeadline(deadlinesCalculator.plus28DaysAt4pmDeadline(
+                LocalDateTime.now()));
+        }
         return AboutToStartOrSubmitCallbackResponse.builder()
-                .data(dataBuilder.build().toMap(objectMapper))
-                .build();
+            .data(dataBuilder.build().toMap(objectMapper))
+            .build();
     }
 
 }
